Initial commit
This is not working yet.

Managed to get all services up and running, but Tomcat and the workers
don't seem to start properly and return errors, but if I start them
manually it seems to be fine...

Networking between services seems to be working.

Can access the web app but when archiving a localhost deposit it fails in the middle of the process...